Move `simple_properties_identifiers` to linter
