+++ src/extension/alterschema/linter/simple_properties_identifiers.h
