+++ test/alterschema/alterschema_canonicalize_2020_12_test.cc
@@ -1322,48 +1322,3 @@ TEST(AlterSchema_canonicalize_2020_12,
 
   EXPECT_EQ(document, expected);
 }
-
-TEST(AlterSchema_canonicalize_2020_12,
-     simple_properties_identifiers_skip_metaschema) {
-  const sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2020-12/schema",
-    "$id": "https://example.com/my-metaschema",
-    "$vocabulary": {
-      "https://json-schema.org/draft/2020-12/vocab/core": true,
-      "https://json-schema.org/draft/2020-12/vocab/applicator": true,
-      "https://json-schema.org/draft/2020-12/vocab/validation": true
-    },
-    "type": "object",
-    "minProperties": 0,
-    "properties": {
-      "foo-bar": { "type": "string", "minLength": 0 }
-    }
-  })JSON");
-
-  CANONICALIZE_WITHOUT_FIX(document, result, traces);
-
-  EXPECT_TRUE(result.first);
-  EXPECT_EQ(traces.size(), 0);
-}
-
-TEST(AlterSchema_canonicalize_2020_12,
-     simple_properties_identifiers_applies_non_meta) {
-  const sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2020-12/schema",
-    "$id": "https://example.com/my-schema",
-    "type": "object",
-    "minProperties": 0,
-    "properties": {
-      "foo-bar": { "type": "string", "minLength": 0 }
-    }
-  })JSON");
-
-  CANONICALIZE_WITHOUT_FIX(document, result, traces);
-
-  EXPECT_FALSE(result.first);
-  EXPECT_EQ(traces.size(), 1);
-  EXPECT_LINT_TRACE(traces, 0, "", "simple_properties_identifiers",
-                    "Set `properties` to identifier names that can be easily "
-                    "mapped to programming languages (matching "
-                    "[A-Za-z_][A-Za-z0-9_]*)");
-}
